removed methods from table class
I removed all the methods from the table and PokerInfo classes and instead put all the logic in the Game class.
This makes it more efficient to send the table to the players and relies more on the Game class to treat everything related to the logic.
Table and PlayerInfo now just store data in public variables.