Created a winning probability simulator
Created a function that computes the winning probability of a player, given the number of currently active players, the state of the table (the public cards) and some number of simulations. Has accuracy roughly 1/sqrt(num_simulations), being a Monte-Carlo algorithm. Looking into going fully deterministic into smaller cases (after the flop stage for example). A lot of prints for debugging purposes that I will get rid of.