local class of player gone / removed call, raise, fold
Removed the class of PokerPlayerLocal, only the PokerPlayer class remains as the other one was pretty useless.

Also removed the call, fold, raise functions from the PokerPlayer class
this will be handled by the Action function instead which will return a value.
The class will not directly communicate with the server but through a different object that will call Action when needed