Changed the bot to integrate action function
Changed the bot accordingly to the action function that I found in some branch. This is not commited on main, so it shows erros, nor it is implemented from what I saw, nor commented to get how it should work. I do not know if this will work then. There is no visible atribute for the number of players still in game at some stage.